+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Criado o componente `MatchHistoryStatsAccordion` (https://github.com/3C-gg/reload-frontend/issues/154).
+- Tema customizado para component Progress do Chakra UI.
+- Criado componente `HeaderProfile` (https://github.com/3C-gg/reload-frontend/issues/153).
 - Adicionado lib `react-countup` para animar um número de 0 a x.
 - Criado componente `LevelProgressBar` (https://github.com/3C-gg/reload-frontend/issues/137).
 - Criado `MatchStatsTable` componente (https://github.com/3C-gg/reload-frontend/issues/138).